# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GreenGenius API
  description: Backend for the GreenGenius AI plant pot.
  version: 1.0.0
paths:
...
  /v1/pots:
    get:
      tags:
      - pots
      summary: List Pots
      operationId: list_pots_v1_pots_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PotOut'
                title: Response List Pots V1 Pots Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - pots
      summary: Create Pot
      operationId: create_pot_v1_pots_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PotCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pots/{pot_id}:
    get:
      tags:
      - pots
      summary: Get Pot
      operationId: get_pot_v1_pots__pot_id__get
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - pots
      summary: Update Pot
      operationId: update_pot_v1_pots__pot_id__patch
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PotUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - pots
      summary: Delete Pot
      operationId: delete_pot_v1_pots__pot_id__delete
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pots/{pot_id}/latest:
    get:
      tags:
      - pots
      summary: Get Latest
      operationId: get_latest_v1_pots__pot_id__latest_get
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pots/{pot_id}/readings:
    get:
      tags:
      - pots
      summary: Get Readings
      description: 'Bucketed time series.


        Aggregation happens in SQL, not Python: a 90-day window at 60 s sampling is

        ~130k rows, and shipping those to the API process to average them would be

        both slow and pointless.


        On Postgres this targets Timescale''s `time_bucket`; the SQLite path uses an

        epoch-division expression so local development produces the same shape.'
      operationId: get_readings_v1_pots__pot_id__readings_get
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: bucket
        in: query
        required: false
        schema:
          enum:
          - 5m
          - 1h
          - 1d
          type: string
          default: 1h
          title: Bucket
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 2160
          minimum: 1
          default: 24
          title: Hours
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LatestOut:
      properties:
        pot_id:
          type: string
          title: Pot Id
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        online:
          type: boolean
          title: Online
        reading:
          anyOf:
          - $ref: '#/components/schemas/ReadingOut'
          - type: 'null'
        health:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Health
      type: object
      required:
      - pot_id
      - device_id
      - online
      - reading
      - health
      title: LatestOut
...
    PotCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      type: object
      required:
      - name
      title: PotCreate
    PotOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Url
        identify_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Identify Confidence
        auto_water_enabled:
          type: boolean
          title: Auto Water Enabled
        auto_water_threshold_pct:
          type: number
          title: Auto Water Threshold Pct
        created_at:
          type: string
          format: date-time
          title: Created At
        species:
          anyOf:
          - $ref: '#/components/schemas/SpeciesOut'
          - type: 'null'
      type: object
      required:
      - id
      - name
      - device_id
      - photo_url
      - identify_confidence
      - auto_water_enabled
      - auto_water_threshold_pct
      - created_at
      title: PotOut
    PotUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
        auto_water_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Auto Water Enabled
        auto_water_threshold_pct:
          anyOf:
          - type: number
            maximum: 100.0
            minimum: 0.0
          - type: 'null'
          title: Auto Water Threshold Pct
      type: object
      title: PotUpdate
    ReadingOut:
      properties:
        time:
          type: string
          format: date-time
          title: Time
        temp_c:
          anyOf:
          - type: number
          - type: 'null'
          title: Temp C
        rh:
          anyOf:
          - type: number
          - type: 'null'
          title: Rh
        soil_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Soil Pct
        lux:
          anyOf:
          - type: number
          - type: 'null'
          title: Lux
        batt_mv:
          anyOf:
          - type: integer
          - type: 'null'
          title: Batt Mv
        flags:
          type: integer
          title: Flags
          default: 0
      type: object
      required:
      - time
      - temp_c
      - rh
      - soil_pct
      - lux
      title: ReadingOut
    SeriesOut:
      properties:
        pot_id:
          type: string
          title: Pot Id
        bucket:
          type: string
          title: Bucket
        points:
          items:
            $ref: '#/components/schemas/SeriesPoint'
          type: array
          title: Points
      type: object
      required:
      - pot_id
      - bucket
      - points
      title: SeriesOut
    SeriesPoint:
      properties:
        bucket:
          type: string
          format: date-time
          title: Bucket
        temp_c:
          anyOf:
          - type: number
          - type: 'null'
          title: Temp C
        rh:
          anyOf:
          - type: number
          - type: 'null'
          title: Rh
        soil_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Soil Pct
        lux:
          anyOf:
          - type: number
          - type: 'null'
          title: Lux
        samples:
          type: integer
          title: Samples
          default: 0
      type: object
      required:
      - bucket
      title: SeriesPoint
    SpeciesOut:
      properties:
        id:
          type: string
          title: Id
        scientific_name:
          type: string
          title: Scientific Name
        common_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Common Name
      type: object
      required:
      - id
      - scientific_name
      - common_name
      title: SpeciesOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...